backend/requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
mysql-connector-python==8.4.0

# Data processing
pandas==2.2.2
matplotlib==3.9.0

# LLM / Vector store
langchain-ollama==0.1.1
langchain-community==0.2.5
langchain-core==0.2.10
faiss-cpu==1.8.0

# HTTP client (needed by FastAPI middleware and health checks)
httpx==0.27.0
